requirements.txt
Django==1.9.6
djangorestframework>=3.3
psycopg2
drf-nested-routers
django-compressor
